Added an S3 bucket and policies to the role assumed by the EC2 instance that allow interactions with it.

+++ reset.sh
@@ -1,4 +1,25 @@
 #!/bin/bash
-rm terraform.tfstate
-rm inventory.ini
-rm connect_to_ec2.sh
+
+# Check if terraform.tfstate exists and remove it if it does
+if [ -f terraform.tfstate ]; then
+    rm terraform.tfstate
+    echo "terraform.tfstate removed"
+else
+    echo "terraform.tfstate does not exist"
+fi
+
+# Check if inventory.ini exists and remove it if it does
+if [ -f inventory.ini ]; then
+    rm inventory.ini
+    echo "inventory.ini removed"
+else
+    echo "inventory.ini does not exist"
+fi
+
+# Check if connect_to_ec2.sh exists and remove it if it does
+if [ -f connect_to_ec2.sh ]; then
+    rm connect_to_ec2.sh
+    echo "connect_to_ec2.sh removed"
+else
+    echo "connect_to_ec2.sh does not exist"
+fi
